move the keyword/mark matching to the "legacy" module
